requirements-test.txt
google_cloud_storage>=1.35
pandas
orjson
matplotlib
rich
python-dotenv
networkx
moz-sql-parser
gcp_storage_emulator
xmltodict
psutil
pyarrow
rsa>=4.7 # not directly required, pinned by Snyk to avoid a vulnerability
minio
